sdk/pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "verdictlens"
version = "0.2.0"
description = "Open-source observability for AI agents — hierarchical tracing, auto-instrumentation, blame analysis."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [{ name = "VerdictLens" }]
keywords = ["ai", "agents", "observability", "tracing", "llm", "openai", "langchain", "debugging"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Debuggers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Typing :: Typed",
]
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]
openai = ["openai>=1.0.0"]
langchain = ["langchain-core>=0.2.0"]
crewai = ["crewai>=0.30.0"]
autogen = ["pyautogen>=0.2.0"]
llamaindex = ["llama-index-core>=0.10.0"]
otel = [
  "opentelemetry-sdk>=1.20.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
]
all = [
  "openai>=1.0.0",
  "langchain-core>=0.2.0",
  "crewai>=0.30.0",
  "pyautogen>=0.2.0",
  "llama-index-core>=0.10.0",
  "opentelemetry-sdk>=1.20.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
]

[project.scripts]
verdictlens-ci = "verdictlens.cli:main"

[project.urls]
Homepage = "https://github.com/verdictlens/verdictlens"
Documentation = "https://github.com/verdictlens/verdictlens#readme"
Repository = "https://github.com/verdictlens/verdictlens"
"Bug Tracker" = "https://github.com/verdictlens/verdictlens/issues"

[tool.hatch.build.targets.wheel]
packages = ["verdictlens"]

[tool.hatch.build.targets.sdist]
include = ["verdictlens", "README.md", "pyproject.toml"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
